renaming synth_temp_network module
