+++ src/flowstab/synth_temp_network.py
